working in class clean
