+++ UnitTestGenerator/src/main/java/com/unitTestGenerator/interfaces/IPorjectName.java
@@ -2,11 +2,8 @@ package com.unitTestGenerator.interfaces;
 
 import com.unitTestGenerator.core.ProjectHolder;
 import com.unitTestGenerator.ioc.ContextIOC;
-import com.unitTestGenerator.test.ioc.IocConected;
 import com.unitTestGenerator.util.IConstantModel;
 
-import java.util.ArrayList;
-import java.util.List;
 
 public interface IPorjectName {
