+++ UnitTestGenerator/src/main/java/com/unitTestGenerator/menus/SavedProjectMenu.java
@@ -11,7 +11,6 @@ import com.unitTestGenerator.persistence.repositories.IDaoService;
 import com.unitTestGenerator.pojos.Project;
 import com.unitTestGenerator.printers.IPrintService;
 import com.unitTestGenerator.printers.PrintProjectAnalyzers;
-import com.unitTestGenerator.test.AppTestDatabase;
 import com.unitTestGenerator.util.ObjectTransformer;
 
 import java.util.ArrayList;
@@ -33,7 +32,7 @@ public class SavedProjectMenu implements IPrintService {
         this.objectTransformer = objectTransformer;
         this.projectHolder = projectHolder;
         this.analizerProjectService = analizerProjectService;
-        this.dao =  ContextIOC.getInstance(AppTestDatabase.class).getClassInstance(DataDAOServicesImplement.class);
+        this.dao =  ContextIOC.getInstance().getClassInstance(DataDAOServicesImplement.class);
     }
 
     public Project started(Scanner scanner){
