--- UnitTestGenerator/src/main/java/com/unitTestGenerator/test/ioc/Servicio.java
@@ -1,5 +1,0 @@
-package com.unitTestGenerator.test.ioc;
-
-public interface Servicio {
-    void ejecutar();
-}
