--- UnitTestGenerator/src/main/java/com/unitTestGenerator/test/testIOC.java
@@ -1,18 +1,0 @@
-package com.unitTestGenerator.test;
-
-
-import com.unitTestGenerator.ioc.*;
-import com.unitTestGenerator.ioc.anotations.EndebleIOC;
-import com.unitTestGenerator.test.ioc.IocConected;
-
-@EndebleIOC("com.unitTestGenerator.test.ioc")
-public class testIOC {
-
-    public static void main(String[] args) {
-        ContextIOC.getInstance(testIOC.class).getClassInstance(IocConected.class).ejecute();
-    }
-
-}
-
-
-
